save changes

+++ ObjectsExchange/Controllers/ClientNodeController.cs
@@ -1,8 +1,10 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
 using ObjectsExchange.Client.Models;
 using ObjectsExchange.Data;
+using ObjectsExchange.Services;
 using Radzen;
 using Sabatex.ObjectsExchange.Models;
 using Sabatex.RadzenBlazor;
@@ -14,7 +16,7 @@ namespace ObjectsExchange.Controllers;
 public class ClientNodeController : BaseController<ObjectsExchange.Client.Models.ClientNode>
 {
 
-    public ClientNodeController(ObjectsExchangeDbContext context,ILogger<ClientNodeController> logger) : base(context,logger)
+    public ClientNodeController(ObjectsExchangeDbContext context,ILogger<ClientNodeController> logger,ClientManager clientManager) : base(context,logger, clientManager)
     {
 
     }
@@ -25,9 +27,8 @@ public class ClientNodeController : BaseController<ObjectsExchange.Client.Models
         var node = await context.ClientNodes.Include(c=>c.Client).SingleOrDefaultAsync(s=>s.Id ==id);
         if (node == null) { return NotFound(); }
 
-        if (User.IsInRole("Administrator") || node.Client.UserId == UserId)
+        if (User.IsInRole("Administrator") || (node.Client !=null && node.Client.UserId == UserId))
         {
-            context.QueryObjects.Where(s => s.Destination == id).ExecuteDelete();
             context.ObjectExchanges.Where(s=>s.Destination == id).ExecuteDelete();
             return new NoContentResult();
         }
@@ -45,7 +46,7 @@ public class ClientNodeController : BaseController<ObjectsExchange.Client.Models
     {
         var q = base.OnBeforeGet(query, queryParams);
         if (!User.IsInRole("Administrator"))
-            q = q.Where(s => s.Client.UserId == UserId);
+            q = q.Where(s => s.Client!=null &&  s.Client.UserId == UserId);
         
         q=q.Select(x =>new ClientNode
         {
@@ -53,14 +54,42 @@ public class ClientNodeController : BaseController<ObjectsExchange.Client.Models
             ClientId = x.ClientId,
             Name = x.Name,
             Counter = x.Counter,
-            QueriesCount = context.QueryObjects.Where(n => n.Destination == x.Id).Count(),
             ObjectsCount = context.ObjectExchanges.Where(n => n.Destination == x.Id).Count()
         });
 
         return q;
     }
 
+    protected override async Task OnBeforeAddItemToDatabase(ClientNode item)
+    {
+        await base.OnBeforeAddItemToDatabase(item);
+        item.Password = clientManager.GetHashString(item.Password);
+    }
 
+    protected override async Task OnBeforeUpdateAsync(ClientNode item, ClientNode update)
+    {
+        await base.OnBeforeUpdateAsync(item,update);
+        if (string.IsNullOrWhiteSpace(update.Password))
+            update.Password = item.Password;
+        else
+            update.Password = clientManager.GetHashString(update.Password);
+    } 
+
+    protected override async Task OnAfterGetById(ClientNode item, Guid id)
+    {
+        item.Password = string.Empty;
+        await Task.Yield();
+    }
+
+    protected override async Task<bool> CheckAccess(ClientNode item, ClientNode? update)
+    {
+        if (update != null)
+        {
+            if (item.ClientId != update.ClientId) return false;
+        }
+        var client = await context.Clients.SingleAsync(s=>s.Id == item.ClientId);
+        return client.UserId == UserId;
+    }
 }
